chore: move deepMerge to utils package
Small refactor to use this function later in the kubernetes/lib
package.
