+++ lib/utils/src/index.ts
@@ -27,6 +27,7 @@ export * from "./sagas";
 export * from "./errors";
 export * from "./diffutils";
 export * from "./pubkey";
+export * from "./merge";
 export { BUILD_INFO } from "./buildinfo";
 
 export {
