+++ lib/utils/src/merge.ts
@@ -1,0 +1,98 @@
+/**
+ * Copyright 2021 Opstrace, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+//
+// Code below was extracted  from
+// https://gist.github.com/mir4ef/c172583bdb968951d9e57fb50d44c3f7 to do deep
+// object merging with typescript.
+//
+interface IIsObject {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  (item: any): boolean;
+}
+
+interface IObject {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  [key: string]: any;
+}
+
+interface IDeepMerge {
+  (target: IObject, ...sources: Array<IObject>): IObject;
+}
+
+/**
+ * @description Method to check if an item is an object. Date and Function are considered
+ * an object, so if you need to exclude those, please update the method accordingly.
+ * @param item - The item that needs to be checked
+ * @return {Boolean} Whether or not @item is an object
+ */
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const isObject: IIsObject = (item: any): boolean => {
+  return item === Object(item) && !Array.isArray(item);
+};
+
+/**
+ * @description Method to perform a deep merge of objects
+ * @param {Object} target - The targeted object that needs to be merged with the supplied @sources
+ * @param {Array<Object>} sources - The source(s) that will be used to update the @target object
+ * @return {Object} The final merged object
+ */
+export const deepMerge: IDeepMerge = (
+  target: IObject,
+  ...sources: Array<IObject>
+): IObject => {
+  // return the target if no sources passed
+  if (!sources.length) {
+    return target;
+  }
+
+  const result: IObject = target;
+
+  if (isObject(result)) {
+    const len: number = sources.length;
+
+    for (let i = 0; i < len; i += 1) {
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const elm: any = sources[i];
+
+      if (isObject(elm)) {
+        for (const key in elm) {
+          // eslint-disable-next-line no-prototype-builtins
+          if (elm.hasOwnProperty(key)) {
+            if (isObject(elm[key])) {
+              if (!result[key] || !isObject(result[key])) {
+                result[key] = {};
+              }
+              deepMerge(result[key], elm[key]);
+            } else {
+              if (Array.isArray(result[key]) && Array.isArray(elm[key])) {
+                // concatenate the two arrays and remove any duplicate primitive values
+                result[key] = Array.from(new Set(result[key].concat(elm[key])));
+              } else {
+                result[key] = elm[key];
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  return result;
+};
+//
+// End of gist code.
+//
